requirements.txt
openai>=1.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
python-pptx>=0.6.23
beautifulsoup4>=4.12.0
pyyaml>=6.0
markdown>=3.5
rich>=13.0
python-dotenv>=1.0.0
